🧠 Academic Club (DSL) Projects/25-2_DSL_Modeling_NLP2_HospitalAgent/25-2_DSL_Modeling_NLP2_HospitalAgent/requirements.txt
# Medical Reservation Agent - Requirements

# FastAPI 및 웹 서버
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0

# 데이터 처리
pandas>=1.5.0
numpy>=1.20.0

# 머신러닝 및 임베딩
sentence-transformers>=2.0.0
scikit-learn>=1.0.0
torch>=1.12.0

# Google AI (Gemini)
google-generativeai>=0.3.0

# 기타 유틸리티
python-dotenv>=0.19.0
openpyxl>=3.0.0  # Excel 파일 읽기용

# 개발 및 테스트
pytest>=7.0.0
pytest-asyncio>=0.20.0
sentence-transformers


# LangGraph
langgraph>=0.0.40
langchain>=0.1.0
langchain-core>=0.1.0

# Supabase MCP
mcp>=1.0.0
supabase>=2.0.0
postgrest>=0.13.0

# Supabase REST API
requests>=2.25.0
aiohttp

# Tavily 검색
tavily-python>=0.3.0

# LangGraph CLI
langgraph-cli>=0.0.40

# Jupyter Notebook for testing and visualization
jupyter>=1.0.0
ipykernel>=6.0.0
matplotlib>=3.5.0
networkx>=2.6.0

# LangChain MCP Adapters (MCP 통합)
langchain-mcp-adapters>=0.1.0

# LangSmith 연동
langsmith>=0.1.0

# OpenAI
openai>=1.0.0
langchain-openai>=0.1.0

websockets
attrs
